Reran time data
Here are the time data and statistical descriptions and the correlation matrix that goes with it. I have ended up filtering out the least amount of data with the tradeoffs between mean_patient_count and mean_triage_time, where, if I trim the start of the scenario to something closer to approximating the freeform teleport, I have to filter out 15 scenes with low patient counts. Some bullet points:
This zeroes out the mean_time_to_hemorrhage_control_per_patient for the scene, per patient, as we trim off the action ticks at the beginning of the scene so that the first action tick ends up as the participant engaging the only hemorrhaging, non-dead patient.
sum_action_count is the sum of the participant's player actions of strictly the PULSE_TAKEN, TOOL_APPLIED, and TAG_APPLIED action types
sum_discarded_count is the sum of the participant's "indecision metric" of TAG_DISCARDED and TOOL_DISCARDED counts
mean_tool_indecision_time is the mean of the participant's time between first-in-sequence TOOL_HOVER and last-in-sequence TOOL_SELECTED